python/arch_diagram_harvester/requirements.txt
google-cloud-aiplatform
google-cloud-storage
google-cloud-vision
google-cloud-firestore